Changed reducers to TS. Added Store type
