--- src/redux/reducers/index.js
@@ -1,8 +1,0 @@
-import { combineReducers } from 'redux';
-import isSSR from './isSSR';
-
-const rootReducer = combineReducers({
-  isSSR,
-});
-
-export default rootReducer;
